Stop misusing the Default trait; bump version number (breaking)
This is a breaking change.

Default::default() no longer returns the content of the `.env`
file, returning instead the environment variables exported by the
operative system only.

Dotenv::dotenv() returns the content of the `.env` variable.

This is more idiomatic on its use of the Default trait, and allows
it to be used for lazy error handling through `unwrap_or_default`.