Fix directory when using Cargo
When a project is run via Cargo, `current_dir()` returns the
directory `cargo run` was called from, essentially an arbitrary
location. This is confirmed working as intended.

Previously, `dotenv()` would check `current_dir()` and error if
it could not find an `.env` file there. This change makes it so
that it will also check the directory Cargo loaded `Cargo.toml`
from before returning an error.

Projects that do not use Cargo are not affected by this change.