Change from_filename() to check parent directories
This change makes it so `from_filename()` (and consequently
`dotenv()`) first attempts to load from the working directory.
Then if the file is not found, it tries again in the parent,
repeating until the file is found or the top-level directory is
reached.  This is a superset of the directories checked in
previous versions.

Closes #18