Stop parsing trailing args without need for --
Previously, clap would try and parse arguments from the entire arg list,
even though we only care about the options we've explicitly specified.
This means that an invocation like

    dotenv -f .env.development sh -c 'echo $DATABASE_URL'

would incorrectly exit with an error, because '-c' is an unrecognised
argument.

Following the suggested approach in
https://github.com/kbknapp/clap-rs/issues/484#issuecomment-211126542, we
use the `AllowExternalSubcommands` setting to treat everything after the
option list as a subcommand, letting us pass through the remaining
arguments to the command untouched. The tradeoff is that we need to
provide a custom error message when no command is supplied.