Fix comments reggression
This PR fixes #12

Added a WhiteSpace ParseState and check for a comment after a
whitespace. Also return the current position of the iter to correctly
strip the comments of from the line to prevent quoted comments or
non-whitespace surrounded hash chars.

Also updated the tests to match these cases.
And added an example code to easily show the output.